Changed the whole predictValue function

+++ src/arch/x86/superop/trace_based_graph.hh
@@ -39,7 +39,7 @@ class TraceBasedGraph : public SimObject
     bool usingControlTracking = false;
     bool usingCCTracking = false;
 
-    void predictValue(Addr addr, unsigned uopAddr, int64_t value, int8_t confidence, unsigned latency);
+    void predictValue(Addr addr, uint16_t uopAddr, uint64_t value, unsigned confidence, unsigned latency);
 
     bool updateSpecTrace(SpecTrace &trace, bool& isDeadCode, bool propagated);
